Rename `box_.rs` to `fragment.rs`
